requirements.txt
pandas>=2.0.0
pyupbit>=0.2.34
python-dotenv>=1.0.0
requests>=2.31.0
